requirements.txt
# Dépendances du projet IoT - Supervision Intelligente
# =====================================================
# Installation : pip install -r requirements.txt

# Configuration
python-dotenv>=1.0.0

# Communication MQTT
paho-mqtt>=1.6.1

# Traitement des données
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning - Détection d'anomalies
scikit-learn>=1.3.0

# Visualisation
matplotlib>=3.7.0
plotly>=5.15.0

# Dashboard Web
streamlit>=1.28.0

# Utilitaires
python-dotenv>=1.0.0

# =====================================================
# DÉPENDANCES CLOUD (Optionnel)
# =====================================================

# MongoDB Atlas
pymongo>=4.5.0

# Certificats SSL
certifi>=2024.0.0

